Deploy vllm service with known ports
kserve, may create a headless service depending on cluster
configuration. If serviceClusterIPNone=true, the service will be
headless and we'll need to reach vllm using port 8080, otherwise it will
have an IP and we'll need to reach vllm using port 80.
In order to work around this cluster configuration, this PR adds a
service named $key-vllm, that will listen on port 80 and use targetPort
8080.

Signed-off-by: Yuval Turgeman <yturgema@redhat.com>